Modificando arquivos para suporte a psr-4
