+++ .gitignore
@@ -7,3 +7,16 @@ nbdist/
 nbactions.xml
 nb-configuration.xml
 .nb-gradle/
+
+
+# Created by https://www.gitignore.io/api/composer
+
+### Composer ###
+composer.phar
+/vendor/
+
+# Commit your application's lock file http://getcomposer.org/doc/01-basic-usage.md#composer-lock-the-lock-file
+# You may choose to ignore a library lock file http://getcomposer.org/doc/02-libraries.md#lock-file
+# composer.lock
+
+# End of https://www.gitignore.io/api/composer
